+++ src/main/java/com/nickesqueda/socialmediademo/service/PostService.java
@@ -1,12 +1,13 @@
 package com.nickesqueda.socialmediademo.service;
 
 import com.nickesqueda.socialmediademo.dto.PostDto;
-import com.nickesqueda.socialmediademo.entity.Comment;
 import com.nickesqueda.socialmediademo.entity.Post;
 import com.nickesqueda.socialmediademo.entity.UserEntity;
+import com.nickesqueda.socialmediademo.exception.UnauthorizedOperationException;
 import com.nickesqueda.socialmediademo.mapper.PostMapper;
 import com.nickesqueda.socialmediademo.repository.PostRepository;
 import com.nickesqueda.socialmediademo.repository.UserRepository;
+import com.nickesqueda.socialmediademo.security.AuthUtils;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -18,10 +19,13 @@ import java.util.Optional;
 public class PostService {
   private final PostRepository postRepository;
   private final UserRepository userRepository;
+  private final AuthUtils authUtils;
 
-  public PostService(PostRepository postRepository, UserRepository userRepository) {
+  public PostService(
+      PostRepository postRepository, UserRepository userRepository, AuthUtils authUtils) {
     this.postRepository = postRepository;
     this.userRepository = userRepository;
+    this.authUtils = authUtils;
   }
 
   public void createPost(int userId, PostDto postDto) {
@@ -31,10 +35,15 @@ public class PostService {
             () ->
                 new EntityNotFoundException(
                     "Cannot create post - parent userId " + userId + " not found."));
+    UserEntity currentUser = authUtils.getCurrentlyAuthenticatedUserEntity();
 
-    Post postEntity = PostMapper.toEntity(postDto);
-    postEntity.setUser(userEntity);
-    postRepository.save(postEntity);
+    if (currentUser.equals(userEntity)) {
+      Post postEntity = PostMapper.toEntity(postDto);
+      postEntity.setUser(userEntity);
+      postRepository.save(postEntity);
+    } else {
+      throw new UnauthorizedOperationException("User is not authorized to perform this action");
+    }
   }
 
   public PostDto getPost(int id) {
@@ -44,21 +53,35 @@ public class PostService {
   }
 
   public PostDto updatePost(int postId, PostDto updatedPost) {
-    // TODO: optimize - remove additional find query.
     Optional<Post> postEntityOptional = postRepository.findById(postId);
     Post postEntity =
         postEntityOptional.orElseThrow(
             () ->
                 new EntityNotFoundException(
                     "Cannot update post - postId " + postId + " not found."));
+    UserEntity currentUser = authUtils.getCurrentlyAuthenticatedUserEntity();
 
-    postEntity.setContent(updatedPost.getContent());
-    postRepository.save(postEntity); // TODO: is .save() needed?
-    return PostMapper.toDto(postEntity);
+    if (currentUser.equals(postEntity.getUser())) {
+      postEntity.setContent(updatedPost.getContent());
+      postRepository.save(postEntity);
+      return PostMapper.toDto(postEntity);
+    } else {
+      throw new UnauthorizedOperationException("User is not authorized to perform this action");
+    }
   }
 
   public void deletePost(int postId) {
-    postRepository.deleteById(postId);
+    Optional<Post> postOptional = postRepository.findById(postId);
+    Post postEntity =
+        postOptional.orElseThrow(
+            () -> new EntityNotFoundException("postId " + postId + " not found."));
+    UserEntity currentUser = authUtils.getCurrentlyAuthenticatedUserEntity();
+
+    if (currentUser.equals(postEntity.getUser())) {
+      postRepository.deleteById(postId);
+    } else {
+      throw new UnauthorizedOperationException("User is not authorized to perform this action");
+    }
   }
 
   public List<PostDto> getUsersPosts(int userId) {
@@ -68,6 +91,11 @@ public class PostService {
 
   @Transactional
   public void deleteUsersPosts(int userId) {
-    postRepository.deleteByUserId(userId);
+    UserEntity currentUser = authUtils.getCurrentlyAuthenticatedUserEntity();
+    if (currentUser.getId() == userId) {
+      postRepository.deleteByUserId(userId);
+    } else {
+      throw new UnauthorizedOperationException("User is not authorized to perform this action");
+    }
   }
 }
